r4.1 clean compile
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@4768 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/floats.c
@@ -24,7 +24,7 @@
 #define FPU_OVERFLOW      _EM_OVERFLOW
 #define FPU_UNDERFLOW     _EM_UNDERFLOW
 #define FPU_PRECISIONLOST _EM_INEXACT
-#define EnableFPUexceptions(mask) _controlfp(~(mask), _MCW_EM)
+#define EnableFPUexceptions(mask) _controlfp((unsigned int)~(mask), _MCW_EM)
 #endif /* nti_ms */
 
 #ifdef nti_bor
